lxc: add has_image() helper and use yaml BaseLoader
The image list data is a little complex for a caller to simply
check if an image is available.  This adds has_image() helper
to do that check.

While parsing image data, LXD tends to trip up Yaml's Loader
when parsing timestamps into datetime format. Use the BaseLoader
instead to avoid parsing timestamps (which may be invalid) and
simply keep them as strings.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>